chore: optimize plugin intergrate
